+++ packages/easy-email-editor/package.json
@@ -69,7 +69,6 @@
   },
   "peerDependencies": {
     "easy-email-core": "^4.3.6",
-    "easy-email-extensions": "^4.16.2",
     "mjml-browser": "^4.10.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
