+++ packages/easy-email-editor/tsconfig.json
@@ -26,7 +26,6 @@
       "@core": ["../easy-email-core/src"],
       "@core/*": ["../easy-email-core/src/*"],
       "easy-email-core": ["../easy-email-core"],
-      "easy-email-extensions": ["../easy-email-extensions/src/index.tsx"],
       "react": ["./node_modules/@types/react"]
     }
   },
