+++ packages/easy-email-extensions/src/components/Form/RichTextToolBar/components/Tools/index.tsx
@@ -1,2 +1,1 @@
 export { Tools } from './Tools';
-export { AvailableTools } from './Tools';
