+++ packages/easy-email-extensions/src/components/Form/RichTextToolBar/index.tsx
@@ -1,2 +1,1 @@
 export { RichTextToolBar } from './RichTextToolBar';
-export { AvailableTools } from './components/Tools';
