+++ packages/easy-email-extensions/src/components/Form/index.tsx
@@ -27,7 +27,6 @@ import { InputSearchProps } from '@arco-design/web-react/es/Input';
 import { ColorPickerField } from './ColorPickerField';
 
 export { RichTextField } from './RichTextField';
-export { AvailableTools } from './RichTextToolBar';
 
 export const TextField = enhancer<InputProps>(Input, value => value);
